requirements.txt
streamlit
st-supabase-connection
supabase
requests
huggingface_hub
Pillow